% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_dir}
\alias{clean_dir}
\title{Clean a directory if it exists
This function removes files and/or sub-directory within the
directory if exists.}
\usage{
clean_dir(folder_path)
}
\arguments{
\item{folder_path}{A character string with the path of the directory to be
cleaned.}
}
\value{
NULL. Only for side effects.
}
\description{
Clean a directory if it exists
This function removes files and/or sub-directory within the
directory if exists.
}
\keyword{internal}
