% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_na}
\alias{count_na}
\title{Count the amount or percentage of \code{NA} in a table by column}
\usage{
count_na(dataset, percent = FALSE)
}
\arguments{
\item{dataset}{A dataframe containing rainfall data from different gauges in
its columns.}

\item{percent}{logical, controls whether to calculate the amount or
percentage of \code{NA}.}
}
\value{
A dataframe.
}
\description{
Count the amount or percentage of \code{NA} in a table by column
}
