% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwswatinput-package.R
\docType{package}
\name{cwswatinput-package}
\alias{cwswatinput-package}
\alias{cwswatinput}
\title{cwswatinput: Climate & Weater SWAT Input.}
\description{
cwswatinput (Climate & Weater SWAT input) is an open-source R package for
preparing weather and climate data from different sources for input in the
Soil & Water Assessment Tool (SWAT). Currently two blocks of processing
routines are implemented, one for the pre-processing of NetCDF and tif raster
files as made available from a increasing number of data-providing
institutions around the globe and a second one for the upscaling of physical
station data by interpolation methods.
}
\author{
Reginal Exavier \email{reginalexavier@rocketmail.com}, Peter Zeilhofer \email{zeilhoferpeter@gmail.com}
}
