% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_aggregations.R
\name{daily_last_value}
\alias{daily_last_value}
\title{Create a daily dataset with the last observed hourly value}
\usage{
daily_last_value(
  folder_in,
  folder_out,
  col_name = "20020101",
  from = "2002-01-01 00",
  to = "2021-05-31 23",
  pattern = ".txt$"
)
}
\arguments{
\item{folder_in}{Path of the input files}

\item{folder_out}{Path where to save the transformed files}

\item{col_name}{The column name for the tables on the output. Usually, the
first date of the time series.}

\item{from}{The first date of the series, including the hour part.}

\item{to}{The last date of the series, including the hour part.}

\item{pattern}{an optional regular expression. Only file names which match
the regular expression will be returned.}
}
\value{
Files with a daily resolution
}
\description{
This function allows to transform an hourly observed time series to a daily
one. Here the returned value is the last observed. This functionality serves
to series with hourly frequency where the value of each hour represents the
sum accumulated so far, thus, the last value represents the daily sum.
}
