% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{daily_max_min}
\alias{daily_max_min}
\title{Create a daily maximum and minimum from hourly observations}
\usage{
daily_max_min(
  folder_in,
  folder_out,
  col_name = "20020101",
  from = "2002-01-01 00",
  to = "2021-05-31 23",
  pattern = ".txt$"
)
}
\arguments{
\item{folder_in}{Path of the input files}

\item{folder_out}{Path where to save the transformed files}

\item{col_name}{The column name for the tables on the output. Usually, the
first date of the time series.}

\item{from}{The first date of the series, including the hour part.}

\item{to}{The last date of the series, including the hour part.}

\item{pattern}{an optional regular expression. Only file names which match
the regular expression will be returned.}
}
\value{
Files with a daily resolution
}
\description{
Create a daily maximum and minimum from hourly observations
}
