% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_aggregation.R
\name{datacube_aggregation}
\alias{datacube_aggregation}
\title{Create a daily aggregation from an hourly datacube}
\usage{
datacube_aggregation(
  input_path,
  output_filename = "",
  fun = mean,
  cores = 1,
  mode = c("agg_fun", "value_at_hour")[1],
  value_hour = 0,
  date_shift_days = 0,
  drop_first_layer = FALSE,
  ...
)
}
\arguments{
\item{input_path}{Path to the datasetcube}

\item{output_filename}{Path to the output file}

\item{fun}{Function to be applied to the datasetcube (default is mean). The
function must be a function that takes a vector as input and returns a
single value. The main functions to be used are: Sum, Mean, Min, Max, First
and Last. For last, use \code{dplyr::last}. To use customized function say, for
example "min", you could use use the format fun = \(x) min(x).
See \code{\link[terra:tapp]{terra::tapp()}} for more information.}

\item{cores}{Number of cores to use for the aggregation. Default is 1. See
\code{\link[terra:tapp]{terra::tapp()}} for more information.
See \code{\link[terra:tapp]{terra::tapp()}} for more information.}

\item{mode}{The mode of aggregation. The options are \code{agg_fun} or
\code{value_at_hour}. The \code{agg_fun} mode applies \code{fun} to all
layers in each day. The \code{value_at_hour} mode returns the layer
timestamped at \code{value_hour}.}

\item{value_hour}{Integer hour between 0 and 23 used when
\code{mode = "value_at_hour"}. The default is 0, which matches products
whose daily accumulated value is timestamped at 00:00 at the end of the
accumulation period.}

\item{date_shift_days}{Whole number of days added to the output layer dates
when \code{mode = "value_at_hour"}. Use \code{-1} for products whose
00:00 timestamp represents the previous day.}

\item{drop_first_layer}{Logical. If TRUE and
\code{mode = "value_at_hour"}, the first selected layer is removed. This is
useful when the first selected layer represents the day before the
requested period.}

\item{...}{Additional arguments to pass to \code{\link[=names_to_date]{names_to_date()}}}
}
\value{
A raster object with the aggregated data
}
\description{
\code{datacube_aggregation()} aggregates a datacube by a given function. The
function can either apply an aggregation function to each day of the datacube
or select the layer timestamped at a given hour.
}
\seealso{
\code{\link[terra:tapp]{terra::tapp()}}, \code{\link[=names_to_date]{names_to_date()}}, \href{https://confluence.ecmwf.int/display/CKB/ERA5+family+post-processed+daily+statistics+documentation}{ERA5 family post-processed daily statistics documentation} # nolint: line_length_linter
}
