% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_name}
\alias{file_name}
\title{Extract the file name}
\usage{
file_name(path)
}
\arguments{
\item{path}{A character string with the file path.}
}
\value{
A character string with the file name without the extension.
}
\description{
This function extracts the file name without the extension from a file path.
It is an internal function used in the \code{unit_converter} function.
}
\keyword{internal}
