% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{files_to_table}
\alias{files_to_table}
\title{Turns multiple time series files into a single table}
\usage{
files_to_table(
  files_path,
  files_pattern,
  start_date = "1970-12-31",
  end_date = "1980-12-31",
  interval = "day",
  na_value = NA,
  neg_to_zero = FALSE
)
}
\arguments{
\item{files_path}{path where the files are.}

\item{files_pattern}{pattern for the observation/station points name.}

\item{start_date}{Inform the start date of the series in the format  \%Y-\%m-\%d.}

\item{end_date}{Inform the end date of the series in the format  \%Y-\%m-\%d.}

\item{interval}{Inform the interval between two observations. See
the function x}

\item{na_value}{Value encoded as not available, use \code{NA} to leave it the
way it is.}

\item{neg_to_zero}{logical. inform whether negative values should be corrected to zero.}
}
\value{
A \code{dataframe}.
}
\description{
Turns multiple time series files into a single table
}
