% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_gap}
\alias{fill_gap}
\title{A wrapper function for filling gaps in the rainfall time series}
\usage{
fill_gap(dataset, corPeriod = "daily")
}
\arguments{
\item{dataset}{A dataframe with first column the time, the rest columns are
rainfall data of different gauges.}

\item{corPeriod}{A string showing the period used in the correlation
computing, e.g. daily, monthly, yearly.}
}
\value{
A dataframe.
}
\description{
This function is a wrapper for the \code{\link[hyfo]{fillGap}} in the
\code{hyfo} \pkg{hyfo} package. The main idea here for this wrapping function
is to preserve the column names as they are in the dataset input.
}
\seealso{
For more detail about the algorithm used to fill the gaps, please
see \code{\link[hyfo]{fillGap}}.
}
