% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_and_methods.R
\name{input_raster}
\alias{input_raster}
\alias{input_raster,character-method}
\alias{character-raster}
\alias{input_raster,SpatRaster-method}
\alias{SpatRaster}
\title{Input Raster}
\usage{
input_raster(x, ...)

\S4method{input_raster}{character}(x, ...)

\S4method{input_raster}{SpatRaster}(x, ...)
}
\arguments{
\item{x}{path (character) to the file or a SpatRaster object.}

\item{\dots}{additional arguments to \code{terra::\link[terra]{rast}}.}
}
\value{
SpatRaster
}
\description{
Method to load ncdf or tiff file and convert them into a SpatRaster object.
}
\seealso{
\code{terra::rast}
}
