% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_and_methods.R
\name{input_table}
\alias{input_table}
\alias{input_table,character-method}
\alias{character-table}
\alias{input_table,data.table-method}
\alias{data.table}
\alias{input_table,data.frame-method}
\alias{data.frame}
\title{Input Table}
\usage{
input_table(x, ...)

\S4method{input_table}{character}(x, ...)

\S4method{input_table}{data.table}(x, ...)

\S4method{input_table}{data.frame}(x, ...)
}
\arguments{
\item{x}{path (character) to the file or a data.table object.}

\item{\dots}{additional arguments to \code{data.table::fread}.}
}
\value{
data.table
}
\description{
Method to load a table file and convert it into a data.table object.
}
\seealso{
\code{data.table::fread}
}
\keyword{internal}
