% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_and_methods.R
\name{input_vector}
\alias{input_vector}
\alias{input_vector,character-method}
\alias{character-vector}
\alias{input_vector,SpatVector-method}
\alias{SpatVector}
\alias{input_vector,sf-method}
\alias{sf}
\title{Input Vector}
\usage{
input_vector(x, ...)

\S4method{input_vector}{character}(x, ...)

\S4method{input_vector}{SpatVector}(x, ...)

\S4method{input_vector}{sf}(x, ...)
}
\arguments{
\item{x}{path (character) to the file or a SpatVector object.}

\item{\dots}{additional arguments to \code{terra::\link[terra]{vect}}.}
}
\value{
SpatVector
}
\description{
Method to load a vector file and convert it into a SpatVector object.
}
\seealso{
\code{terra::vect}
}
\keyword{internal}
