% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_input.R
\name{layerValues2pixelA}
\alias{layerValues2pixelA}
\title{Series of Pixel Values (Array Format)}
\usage{
layerValues2pixelA(layer_values, tb_name, col_name = "20020101")
}
\arguments{
\item{layer_values}{List. Values extracted by raster}

\item{tb_name}{A vector contain the names for every table created. These names are
in the mainTable}

\item{col_name}{A name for the column of everery swatinput table created. Commonly this
name is the first date of time serie beeing analysed.}
}
\value{
A list of table
}
\description{
With the extracted values by raster layer from the (raster2vec) function, this function
organize these values in the format of swat input, i.e, a time serie for every pixel
of the study area.
}
