% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat_metadata.R
\name{main_input_var}
\alias{main_input_var}
\title{Main table constructor by Variable}
\usage{
main_input_var(study_area, var_name = "temp")
}
\arguments{
\item{study_area}{The object from 'study_area_records'}

\item{var_name}{The name of the variable to be extracted}
}
\value{
A table
}
\description{
Construct a main table needed for the input in SWAT
}
\examples{
study_area <- data.frame(
  ID = 1:2,
  LAT = c(-15.0, -15.5),
  LON = c(-56.0, -56.5),
  ELEVATION = c(100, 120)
)
main_input_var(study_area, var_name = "tmin")
}
