% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_input.R
\name{mainInput_var}
\alias{mainInput_var}
\title{Main table contructor by Variable}
\usage{
mainInput_var(study_area, var_name = "temp")
}
\arguments{
\item{study_area}{The object from 'study_area_records'}

\item{var_name}{The name of the variable to be extracted}
}
\value{
A table
}
\description{
Construct a main table needed for the input in SWAT
}
