% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{names_to_date}
\alias{names_to_date}
\title{Extract the date from the layer names
This function extracts the date from the layer names of a raster object.}
\usage{
names_to_date(
  raster_cube,
  origin = "1970-01-01",
  tz = "UTC",
  regex = ".*=(\\\\d+)"
)
}
\arguments{
\item{raster_cube}{A raster object.}

\item{origin}{A character string with the origin date. The default is "1970-01-01".}

\item{tz}{A character string with the time zone. The default is "UTC".}

\item{regex}{A character string with the regular expression to extract the date.}
}
\value{
A POSIXct object.
}
\description{
Extract the date from the layer names
This function extracts the date from the layer names of a raster object.
}
\keyword{internal}
