% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_input.R
\name{ncdf2raster}
\alias{ncdf2raster}
\title{NetCDF to Raster}
\usage{
ncdf2raster(
  ncdf_file,
  var_name = "my_variable",
  time_step_start = 1,
  time_step_end = 1,
  datetime = "2000-01-01",
  coordinate_rs = sp::CRS("+proj=longlat +datum=WGS84")
)
}
\arguments{
\item{ncdf_file}{A ncdf file to be transformed to raster}

\item{var_name}{The variable to be}

\item{time_step_start}{The time step for the first datetime}

\item{time_step_end}{The time step for the last datetime (1 for unique layer
e -1 for all see ncdf4::ncvar_get)}

\item{datetime}{The datetime for the layer}

\item{coordinate_rs}{The coordinate reference system see sp::CRS()}
}
\value{
A raster
}
\description{
NetCDF to Raster
}
