% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_input.R
\name{ncdf_to_raster}
\alias{ncdf_to_raster}
\title{NetCDF to Raster}
\usage{
ncdf_to_raster(
  ncdf_file,
  var_name,
  time_step_end = NULL,
  coordinate_rs = sp::CRS("+proj=longlat +datum=WGS84")
)
}
\arguments{
\item{ncdf_file}{Character string of the path of the NetCDF file to be
transformed into raster.}

\item{var_name}{Character string of the variable name to be extracted.}

\item{time_step_end}{If not NULL, a numeric value for the final time step. When NULL,
all the time steps are read if exist in the raw NetCDF file.}

\item{coordinate_rs}{Character string of the coordinate reference system, see
\code{\link[sp]{CRS}}.}
}
\value{
A raster
}
\description{
Transformation/convertion of a NetCDF file into a Raster file.
}
