% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{point_to_daily}
\alias{point_to_daily}
\title{Transforms the raw value from point perspective to a daily perspective}
\usage{
point_to_daily(
  my_folder,
  var_pattern = "p-",
  main_pattern = "pcp",
  start_date = "20170301",
  end_date = "20170331",
  interval = "day",
  na_value = -99,
  negatif_number = TRUE,
  prefix = "day_"
)
}
\arguments{
\item{my_folder}{character. The path to the raw files.}

\item{var_pattern}{character. A pattern for the observation/station points
name.}

\item{main_pattern}{character. A pattern for the main file  containing all
the points with ID, NAME, LAT, LONG and ELEVATION.}

\item{start_date}{character. Inform the start date of the serie in the format
yyyymmdd.}

\item{end_date}{character. Inform the end date of the serie in the format
yyyymmdd.}

\item{interval}{charactere. Inform the inteval betwenn two observations. See
the function x}

\item{na_value}{numeric. Value encoded as not available, use \code{NA} for
NA.}

\item{negatif_number}{logical. Inform if negative values should be kept.}

\item{prefix}{character. A prefix for naming the table in the format of
"prefix+date". for more detail.}
}
\value{
A list of table
}
\description{
Having the collected observation from a point perspective, this function
transform the input to a vertical perspective, like a a daily perspective.
}
\examples{
folder <- system.file("extdata/pcp_stations", package = "wcswatin")
test01 <- point_to_daily(my_folder = folder)

}
