% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rh_calculator}
\alias{rh_calculator}
\title{Calculate the Relative Humidity from dewpoint and ambient temperature}
\usage{
rh_calculator(
  folder_dpt,
  folder_tas,
  folder_out,
  col_name = "20020101",
  file_name_output = "rh",
  m_value = 7.591386,
  Tn_value = 240.7263,
  pattern = ".txt$"
)
}
\arguments{
\item{folder_dpt}{Path of the input 2m dewpoint temperature files as
\code{Td}.}

\item{folder_tas}{Path of the input Near-Surface Air Temperature files as
\code{Tambient}.}

\item{folder_out}{Path where to save the transformed files}

\item{col_name}{The column name for the tables on the output. Usually, the
first date of the time series.}

\item{file_name_output}{Character string for the Relative humidity files on
output.}

\item{m_value}{The value for the constant \code{m} \cite{(Vaisala, 2013)}.}

\item{Tn_value}{The value for \code{Tn} (Triple point temperature 273.16 K),
constant \cite{(Vaisala, 2013)}.}

\item{pattern}{an optional regular expression. Only file names which match
the regular expression will be returned.}
}
\value{
Files with the same temporal resolution as the input
}
\description{
This function performs the calculation of a relative humidity with dewpoint and
ambient temperature as input applying the formula: \eqn{RH =
100*10^(m*[(Td/(Td+Tn)) - (Tambient/(Tambient+Tn)]))}. Where \eqn{m} and
\eqn{Tn} are constants \cite{(Vaisala, 2013)}.
}
\references{
\href{https://www.vaisala.com/en}{VAISALA} (2013) HUMIDITY
CONVERSION FORMULAS, Calculation formulas for humidity.
}
