% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rh_creator}
\alias{rh_creator}
\title{Create Relative humidity values from 2m dewpoint temperature and Near-Surface
Air Temperature input}
\usage{
rh_creator(
  folder_dpt,
  folder_tas,
  folder_out,
  col_name = "20020101",
  file_name_output = "rh",
  m_value = 7.591386,
  Tn_value = 240.7263,
  pattern = ".txt$"
)
}
\arguments{
\item{folder_dpt}{Path of the input 2m dewpoint temperature files}

\item{folder_tas}{Path of the input Near-Surface Air Temperature files}

\item{folder_out}{Path where to save the transformed files}

\item{col_name}{The column name for the tables on the output. Usually, the
first date of the time series.}

\item{file_name_output}{Character string for the Relative humidity files on
output.}

\item{m_value}{The value for m. Please see \cite{reference}}

\item{Tn_value}{The value for Tn. Please see \cite{reference}}

\item{pattern}{an optional regular expression. Only file names which match
the regular expression will be returned.}
}
\value{
Files with the same temporal resolution as the input
}
\description{
Create Relative humidity values from 2m dewpoint temperature and Near-Surface
Air Temperature input
}
