% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_input.R
\name{study_area_records}
\alias{study_area_records}
\title{Study Area Records}
\usage{
study_area_records(raster_model, roi, dem)
}
\arguments{
\item{raster_model}{One layer representing where the data wile be extracted}

\item{roi}{A shapefile delimiting the study area}

\item{dem}{An elevation raster for the study area}
}
\value{
A table
}
\description{
This function extracts the records position grid for the study area combining
with the DEM for every point
}
