% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_plot}
\alias{summary_plot}
\title{Plot a summary of the data}
\usage{
summary_plot(
  var_folder,
  sample = 5,
  percent = FALSE,
  from = "2002-01-01",
  to = "2021-05-31",
  x_lab = "Months of observation",
  y_lab = "Vriable name and unit",
  pattern = ".txt$"
)
}
\arguments{
\item{var_folder}{Path of the input files}

\item{sample}{Numeric value, informing the number of files to be used.}

\item{percent}{When TRUE, the values passed on \code{sample} is use as as a
percentage.}

\item{from}{The first date of the series.}

\item{to}{The last date of the series.}

\item{x_lab}{Character. Title for the x, see \code{\link[ggplot2]{labs}}.}

\item{y_lab}{Character. Title for the y, see
\code{\link[ggplot2:labs]{labs}}.}

\item{pattern}{an optional regular expression. Only file names which match
the regular expression will be returned.}
}
\value{
A summary plot
}
\description{
Plot a summary of the data
}
