% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_table}
\alias{summary_table}
\title{Table summary of the data}
\usage{
summary_table(
  var_folder,
  sample = 5,
  percent = FALSE,
  by_month = TRUE,
  from = "2002-01-01",
  to = "2021-05-31",
  pattern = ".txt$"
)
}
\arguments{
\item{var_folder}{Path of the input files}

\item{sample}{Numeric value, informing the number of files to be used. Until
the total amount is informed, the choice of points to be computed is
random.}

\item{percent}{When TRUE, the values passed on \code{sample} is use as as a
percentage.}

\item{by_month}{Either the summary should be done per month or in general.
When true, the parameters \code{from} and \code{to} are ignored.}

\item{from}{The first date of the series when \code{by_month} is \code{TRUE}.
Remembering that when \code{by_month} is \code{FALSE}, this parameter is
ignored.}

\item{to}{The last date of the series when \code{by_month} is
\code{TRUE}.Remembering that when \code{by_month} is \code{FALSE}, this
parameter is ignored.}

\item{pattern}{an optional regular expression. Only file names which match
the regular expression will be returned.}
}
\value{
A summary table.
}
\description{
This function creates a table summary containing the \code{min}, the
\code{max}, the \code{mean}, the \code{sd} \emph{(standard deviation)} and
\code{n} \emph{(number of value)} of daily values observed from several
points on a monthly basis. When the parameter \code{by_month} is \code{FALSE}
the summary return is general, i.e., not on a monthly basis, so the table
contains only one row with the same columns \emph{(min, max, mean, sd, and
n)}. You can choose the amount of points to be randomly computed in the total
point set.
}
