% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{touch_dir}
\alias{touch_dir}
\title{Create a directory if it does not exist}
\usage{
touch_dir(folder_path, return_path = FALSE)
}
\arguments{
\item{folder_path}{A character string with the path of the directory to be
created.}

\item{return_path}{logical. If \code{TRUE}, the function returns the path}
}
\value{
NULL. Only for side effects.
}
\description{
This function creates a directory if it does not exist.
It is an internal function used in the \code{unit_converter} function.
}
\keyword{internal}
