% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_surface_interpolation.R
\name{ts_to_area}
\alias{ts_to_area}
\title{Trend Surface Interpolation into Raster}
\usage{
ts_to_area(my_folder, bassin_limit_path, poly_degree = 2, resolution = 0.01)
}
\arguments{
\item{my_folder}{Folder containing the ts input files.}

\item{bassin_limit_path}{A shapefile containing the bassin limit where the
trend suface function have to be predicted.}

\item{poly_degree}{The degree to be used in the polynomial function for the
trend surface.}

\item{resolution}{The resolution for the output raster in degree.}
}
\value{
A multi-layer \code{SpatRaster} with one layer per input file.
}
\description{
This function make an interpolation whith the trend surface method where the
user have to inform the polynome degree. The interpolation is made over the
tageded points for all the serie on the input.
}
\examples{
ts_to_area(
  my_folder = system.file("extdata/ts_input", package = "wcswatin"),
  bassin_limit_path = system.file("extdata/sl_bassin/sl_bassin_limit.shp",
    package = "wcswatin"
  ),
  poly_degree = 2,
  resolution = 0.5
)
}
