% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_interpolation.R
\name{ts_to_point}
\alias{ts_to_point}
\title{Trend Surface Interpolation into Targeded Points}
\usage{
ts_to_point(my_folder, targeted_points_path, poly_degree = 2)
}
\arguments{
\item{my_folder}{Folder containing the ts input files.}

\item{targeted_points_path}{A shapefile containing the targeted points where the
trend suface function have to predict values.}

\item{poly_degree}{The degree to be used in the polynomial function for the trend surface.}
}
\value{
A list of tibles by day
}
\description{
This function make an interpolation whith the trend surface method where the user
have to inform the polynome degree. The interpolation is made over the tageded points
for all the serie on the input.
}
