% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_metadata.R
\name{var_names}
\alias{var_names}
\title{Shows raster variables present in a NetCDF}
\usage{
var_names(path)
}
\arguments{
\item{path}{Path where the netcdf file is located}
}
\value{
Character vector of raster variable names
}
\description{
Shows raster variables present in a NetCDF
}
