% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_interpolation.R
\name{varMain_creator}
\alias{varMain_creator}
\title{Main table creator for SWAT Input from Trend SUrface Interpolation}
\usage{
varMain_creator(targeted_points_path, var_name = "pcp", col_elev = "Elev")
}
\arguments{
\item{targeted_points_path}{Shapefile path}

\item{var_name}{The variable name}

\item{col_elev}{The column contain the elevation values}
}
\value{
A table
}
\description{
This function is to create the main table for the input table for SWAT.
}
\examples{
varMain_creator(targeted_points_path = system.file("extdata/sl_centroides",
               "Centroide_watershed_grau.shp",
               package = "cwswatinput"))
}
