% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcswatin-package.R
\name{wcswatin-package}
\alias{wcswatin-package}
\alias{wcswatin}
\title{wcswatin: Climate & Weather SWAT Input.}
\description{
wcswatin (Climate & Weather SWAT input) is an open-source R package for
preparing weather and climate data from different sources for input in the
Soil & Water Assessment Tool (SWAT), funded by the Critical Ecosystem
Partnership Fund (CEPF). Currently two blocks of processing routines are
implemented, one for the pre-processing of NetCDF and tif raster files as
made available from a increasing number of data-providing institutions around
the globe and a second one for the upscaling of physical station data by
interpolation methods. For processing all used datasets MUST have geographic
coordinates using WGS 84 as datum.
}
\author{
Reginal Exavier \email{reginalexavier@rocketmail.com}, Peter
Zeilhofer \email{zeilhoferpeter@gmail.com}
}
