% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_aggregations.R
\name{windspeed_calculator}
\alias{windspeed_calculator}
\title{Calculate the wind speed from Eastward and Northward Near-Surface Wind}
\usage{
windspeed_calculator(
  folder_uas,
  folder_vas,
  folder_out,
  col_name = "20020101",
  file_name_output = "ws",
  pattern = ".txt$"
)
}
\arguments{
\item{folder_uas}{Path of the input Eastward Near-Surface Wind files
\emph{(as u component)}.}

\item{folder_vas}{Path of the input Northward Near-Surface Wind files
\emph{(as v component)}.}

\item{folder_out}{Path where to save the transformed files}

\item{col_name}{The column name for the tables on the output. Usually, the
first date of the time series.}

\item{file_name_output}{Character string for the Wind speed files on
output.}

\item{pattern}{an optional regular expression. Only file names which match
the regular expression will be returned.}
}
\value{
Files with the same temporal resolution as the input.
}
\description{
This function performs the calculation of the wind speed from Eastward and
Northward Near-Surface Wind as input applying the formula: \eqn{ws =
\sqrt(u^2 + v^2)}.
}
