% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{windspeed_creator}
\alias{windspeed_creator}
\title{Create wind speed values from Eastward Near-Surface Wind and Northward
Near-Surface Wind input}
\usage{
windspeed_creator(
  folder_uas,
  folder_vas,
  folder_out,
  col_name = "20020101",
  file_name_output = "ws",
  pattern = ".txt$"
)
}
\arguments{
\item{folder_uas}{Path of the input Eastward Near-Surface Wind files}

\item{folder_vas}{Path of the input Northward Near-Surface Wind files}

\item{folder_out}{Path where to save the transformed files}

\item{col_name}{The column name for the tables on the output. Usually, the
first date of the time series.}

\item{file_name_output}{Character string for the Wind speed files on
output.}

\item{pattern}{an optional regular expression. Only file names which match
the regular expression will be returned.}
}
\value{
Files with the same temporal resolution as the input
}
\description{
Create wind speed values from Eastward Near-Surface Wind and Northward
Near-Surface Wind input
}
