% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwswatinput-package.R
\docType{package}
\name{cwswatinput-package}
\alias{cwswatinput-package}
\alias{cwswatinput}
\title{cwswatinput: Climate & Weater SWAT Input.}
\description{
Climate & Weater SWAT input (cwswatinput) is an open-source R package for
preparing weather and climate data for the Soil & Water Assessment Tool (SWAT)
input format.
}
\author{
Reginal Exavier \email{reginalexavier@rocketmail.com}, Peter Zeilhofer \email{zeilhoferpeter@gmail.com}
}
