% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_aggregations.R
\name{daily_aggregation}
\alias{daily_aggregation}
\title{Create a daily aggregation from an hourly dataset}
\usage{
daily_aggregation(
  folder_in,
  folder_out,
  pattern = ".txt$",
  from = "2002-01-01 00",
  to = "2021-05-31 23",
  take_out_first_record = TRUE,
  aggregation_function = mean,
  na.rm = FALSE
)
}
\arguments{
\item{folder_in}{Path of the input files}

\item{folder_out}{Path where to save the transformed files}

\item{pattern}{an optional \code{\link[base:regex]{regular expression}}. Only
file names which match the regular expression will be returned.}

\item{from}{The first date of the series, including the hour part.}

\item{to}{The last date of the series, including the hour part.}

\item{take_out_first_record}{Logical. If TRUE, the first record of the input
file will be removed. This is useful when the first record is the hour 00:00, that
corresponds to the previous day. The length in hour between the from and to must be
the same as the length of the hours in the input files.}

\item{aggregation_function}{The function to use on the hourly groups.}

\item{na.rm}{a logical value indicating whether NA values should be removed
before the computation proceeds.}

\item{col_name}{The column name for the tables on the output. Usually, the
first date of the time series.}
}
\value{
Files with a daily resolution
}
\description{
This function allows to aggregate hourly observations to daily
time series. The function for aggregation can be informed in the
\code{aggregation_function} parameter, this parameter takes a function as
argument. The default function is \code{\link[base]{mean}}, so a daily
average is returned.
}
