% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{daily_aggregation}
\alias{daily_aggregation}
\title{Create a daily aggregation from an hourly dataset}
\usage{
daily_aggregation(
  folder_in,
  folder_out,
  pattern = ".txt$",
  col_name = "20020101",
  from = "2002-01-01 00",
  to = "2021-05-31 23",
  aggregation_function = mean,
  na.rm = FALSE
)
}
\arguments{
\item{folder_in}{Path of the input files}

\item{folder_out}{Path where to save the transformed files}

\item{pattern}{an optional \code{\link[base:regex]{regular expression}}. Only
file names which match the regular expression will be returned.}

\item{col_name}{The column name for the tables on the output. Usually, the
first date of the time series.}

\item{from}{The first date of the series, including the hour part.}

\item{to}{The last date of the series, including the hour part.}

\item{aggregation_function}{The function to use on the hourly groups.}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds}
}
\value{
Files with a daily resolution
}
\description{
Create a daily aggregation from an hourly dataset
}
