% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rst_aggregations.R
\name{datacube_aggregation}
\alias{datacube_aggregation}
\title{Create a daily aggregation from an hourly datacube}
\usage{
datacube_aggregation(
  input_path,
  output_filename = "",
  fun = sum,
  cores = 1,
  ...
)
}
\arguments{
\item{input_path}{Path to the datasetcube}

\item{output_filename}{Path to the output file}

\item{fun}{Function to apply to the datasetcube (default is sum). The function
must be a function that takes a vector as input and returns a single value.
See \code{\link[terra:tapp]{terra::tapp()}} for more information.}

\item{cores}{Number of cores to use for the aggregation. Default is 1. See
\code{\link[terra:tapp]{terra::tapp()}} for more information.
See \code{\link[terra:tapp]{terra::tapp()}} for more information.}

\item{...}{Additional arguments to pass to \code{\link[=names_to_date]{names_to_date()}}}
}
\value{
A raster object with the aggregated data
}
\description{
Aggregate a datacube by a given function. The function will be applied to each
day of the datasetcube.
}
\seealso{
\code{\link[terra:tapp]{terra::tapp()}}, \code{\link[=names_to_date]{names_to_date()}}
}
