% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_input.R
\name{layervalues2pixel}
\alias{layervalues2pixel}
\title{Series of Pixel Values}
\usage{
layervalues2pixel(
  layer_values,
  main_tbl,
  col_name = "20220101",
  inline_output = TRUE,
  path_output = NULL,
  append = FALSE
)
}
\arguments{
\item{layer_values}{List. Values extracted by raster}

\item{main_tbl}{A vector contain the names for every table created. These
names are in the mainTable}

\item{col_name}{A name for the column of every swatinput table created.
Commonly this name is the first date of time serie being analysed.}

\item{inline_output}{Logical. If TRUE, the output will be an inline table.}

\item{path_output}{Path to save the individual tables. One of these
parameters (inline_output or path_output) must be provided.}

\item{append}{Logical. If TRUE, the output will be appended to the existing
file. If FALSE, the output will be overwritten. Choose TRUE if you want to
append the output to an existing file, very useful when you are running
different time periods and want to append the output to the same file.}
}
\value{
A list of table or a set of files in the path_output, one for each
pixel
}
\description{
With the extracted values by raster layer from the (cube2table) function,
this function organize these values in the format of swat input, i.e, a time
serie for every pixel of the study area.
}
