% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_input.R
\name{ncdf_to_raster}
\alias{ncdf_to_raster}
\title{NetCDF to Raster}
\usage{
ncdf_to_raster(
  ncdf_file,
  var_name,
  time_step_start = 1,
  time_step_end = 1,
  datetime = NULL,
  coordinate_rs = sp::CRS("+proj=longlat +datum=WGS84")
)
}
\arguments{
\item{ncdf_file}{Character string of the path of the NetCDF file to be
transformed into raster.}

\item{var_name}{Character string of the variable name to be extracted.}

\item{time_step_start}{Numeric value for the time dimension of the first
datetime. See \code{\link[ncdf4]{ncvar_get}}.}

\item{time_step_end}{Numeric value for the time dimension of the last
datetime (1 for unique layer. and -1 for all see
\code{\link[ncdf4]{ncvar_get}}).}

\item{coordinate_rs}{Character string of the coordinate reference system, see
\code{\link[sp]{CRS}}.}
}
\value{
A raster
}
\description{
Transformation/convertion of a NetCDF file into a Raster file.
}
