% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_input.R
\name{raster2vec}
\alias{raster2vec}
\title{Convert a Raster Layer to a Vector}
\usage{
raster2vec(rasterbrick, study_area)
}
\arguments{
\item{rasterbrick}{The raster where the values have to be extracted}

\item{study_area}{The object from 'study_area_records'}
}
\value{
A named list
}
\description{
Convert a Raster Layer to a Vector
}
