% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_daily_tbl}
\alias{save_daily_tbl}
\title{Save the csv files after transformation to daily form}
\usage{
save_daily_tbl(tbl_list, path)
}
\arguments{
\item{tbl_list}{list. A list, the output of the x function}

\item{path}{The path where the files must be saved}
}
\description{
Save the csv files after transformation to daily form
}
\examples{
temp <- tempdir()
folder <- system.file("extdata/pcp_stations", package = "cwswatinput")
test01 <- point_to_daily(my_folder = folder)
save_daily_tbl(
  tbl_list = test01,
  path = temp
)
unlink(temp, recursive = TRUE)
}
