% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_to_files}
\alias{table_to_files}
\title{Export tables to \code{txt} or \code{csv} files}
\usage{
table_to_files(table, folder_path, first_date, file_extention = "txt")
}
\arguments{
\item{table}{A table \code{data.frame} containing all the observations}

\item{folder_path}{Character string of the folder where the file must be
saved.}

\item{first_date}{Character string of the first date for the time series.
This value is used to renaming the columns on every single file while
saving. The actual name is used as the file names. The suggested format is
\code{\%y\%m\%d}}

\item{file_extention}{Character. \code{txt} or \code{csv}.}
}
\description{
Export tables to \code{txt} or \code{csv} files
}
