% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_from_references}
\alias{tbl_from_references}
\title{Create a table with the values extracted from the reference points}
\usage{
tbl_from_references(raster_file, ref_points, prefix_colname = NULL, ...)
}
\arguments{
\item{raster_file}{Raster* object.}

\item{ref_points}{A table or sf object containing the field station reference
points. This input can be a character string informing the address of a
.txt or .csv file, a data.frame or an sf object. The table must have NAME,
LAT and LON fields/columns.}

\item{prefix_colname}{If not null, a string character to be used to prefix
the original column names.}

\item{...}{further arguments from \code{\link[raster]{extract}}. These
arguments concern only extracting the data in the rasters.}
}
\value{
A table.
}
\description{
This function prepares the data table with extracting values in the same
geographical locations as the field stations. This helps to validate data
downloaded from online platforms with data collected from field stations.
}
