% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_names}
\alias{var_names}
\title{Shows the variables present in a netcdf}
\usage{
var_names(path)
}
\arguments{
\item{path}{Path where the netcdf file is located}
}
\value{
List of variable names
}
\description{
Shows the variables present in a netcdf
}
